+++ Runtime/EventMessageBase.cs
@@ -1,9 +1,0 @@
-namespace DJM.EventManager
-{
-    public struct EventMessageBase
-    {
-        public string Something;
-    }
-}
-
-